Feature/#52 팔로우, 팔로잉한 유저 리스트를 가져오는 기능을 구현한다.

+++ src/main/java/com/dinosaur/foodbowl/domain/post/dto/response/PostFeedResponseDto.java
@@ -19,7 +19,7 @@ public class PostFeedResponseDto {
 
   private final String nickname;
   private final String thumbnailPath;
-  private final int followerCount;
+  private final long followerCount;
   private final List<String> photoPaths;
   private final String storeName;
   private final List<String> categories;
